virt-controller, resources, template: move the memory overhead calculation to resource renderer
